Complete working levels representation of tree
This uses a shared one dimensional array to track the count of nodes per
depth. With each new depth it is initialized once to 0 if uninitialized.
It keeps track of the count of the nodes added to that depth. New nodes
at that depth are then added from the tracked count, not from the
beginning (0) of the levels array, preventing a previous overwriting
of values bug.