Print detected neighbor keys on key drag
The maximum neighbors should always be 2, the left and right. This is a
reasonable assumption because of the way x distance is being calculated.
At most it can have 2 next to it in the x. Any further nodes are not
considered. Then y intersection check ensures they are roundabout at the
same level.