Get left and/or right neighbor if they exist
These are saved in a local object returned by the function. This is
better than too many global variables. An object allows us to return
multiple information at once, in one object, like dynamically return
a left and right neighbor and a boolean for if a neighbor exists at all.