Make snap work within levels and from floatingNodes
Currently when snapping from within levels it always snaps in the
direction of the neighbor if only slightly lifted from same position
otherwise it snaps at the right place otherwise. This behavior can be
kept since it functions fine, and can be desired. Alternatively the
behavior can be looked into and changed. Check that within levels
when right/left neighbor is obtained dragged key splices into the right
place. It should preferably stay in the same place when lifted slightly
and only a right or left neighbor is detected.