Try fix timing of show correct tree button
It broke. Not sure why. It was working since I merged the code. The
merge did not work some of my code was missing from index.html for some
reason so I reverted to a commit where it was working. But then show
correct tree button display broke in timing, it showed when it was not
supposed to when the other questions container shows.